+++ src/app/views/_auth/ide/ide.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, ViewChild } from "@angular/core";
-import { MonacoEditorComponent } from "../../components/monaco-editor/monaco-editor.component";
-import { XtermComponent } from "../../components/xterm/xterm.component";
+import { MonacoEditorComponent } from "../../../components/monaco-editor/monaco-editor.component";
+import { XtermComponent } from "../../../components/xterm/xterm.component";
 
 
 @Component({