+++ src/app/views/forgot-password/forgot-passwordComponent.ts
@@ -2,14 +2,15 @@ import { Component, Inject, OnInit } from "@angular/core";
 import { DOCUMENT } from "@angular/common";
 
 @Component({
-  selector: "app-login",
-  templateUrl: "./login.component.html",
+  selector: "app-forgot-password",
+  templateUrl: "./forgot-password.component.html",
 })
-export class LoginComponent implements OnInit {
+export class ForgotPasswordComponent implements OnInit {
   constructor(@Inject(DOCUMENT) private document: Document) {
   }
 
   ngOnInit(): void {
+    this.document.body.className = "dx-viewport";
     this.document.body.classList.add('login-page')
   }
 }