+++ src/app/views/login/login.component.ts
@@ -1,0 +1,16 @@
+import { Component, Inject, OnInit } from "@angular/core";
+import { DOCUMENT } from "@angular/common";
+
+@Component({
+  selector: "app-login",
+  templateUrl: "./login.component.html",
+})
+export class LoginComponent implements OnInit {
+  constructor(@Inject(DOCUMENT) private document: Document) {
+  }
+
+  ngOnInit(): void {
+    this.document.body.className = "dx-viewport";
+    this.document.body.classList.add('login-page')
+  }
+}