+++ CHANGELOG.md
@@ -1,7 +1,33 @@
-## 1.0.0 (2021-06-07)
+## 1.0.8 (2021-07-15)
+
+* Add adminLTE, add bootstrap, add components navbar, aside, login and index. ([56b656f](https://github.com/nonodev96/THUMDER/commit/56b656f))
+* Add component aside and bootstrap ([cbd34ff](https://github.com/nonodev96/THUMDER/commit/cbd34ff))
+* Admin navbar, modals, and tree viewer. ([61d44ce](https://github.com/nonodev96/THUMDER/commit/61d44ce))
+* change budgets ([b58e759](https://github.com/nonodev96/THUMDER/commit/b58e759))
+* fix modals ([20ca7b5](https://github.com/nonodev96/THUMDER/commit/20ca7b5))
+* IDE, Components, forgot-password, etc... ([f455aea](https://github.com/nonodev96/THUMDER/commit/f455aea))
+* Monaco editor + Notus Angular (tailwind + angular) ([e59727c](https://github.com/nonodev96/THUMDER/commit/e59727c))
+
+
+
+## 1.0.0 (2021-06-11)
 
 * Create create-release.yaml ([6c9164c](https://github.com/nonodev96/THUMDER/commit/6c9164c))
+* github cli ([edebb82](https://github.com/nonodev96/THUMDER/commit/edebb82))
+* github cli v2 ([c1dcdf3](https://github.com/nonodev96/THUMDER/commit/c1dcdf3))
+* github cli v3 ([622d1c3](https://github.com/nonodev96/THUMDER/commit/622d1c3))
+* github cli v4 asset_name ([faf2921](https://github.com/nonodev96/THUMDER/commit/faf2921))
+* github cli v5 asset_name ([7adb3c3](https://github.com/nonodev96/THUMDER/commit/7adb3c3))
+* github cli v6 artifactName ([c01bd3a](https://github.com/nonodev96/THUMDER/commit/c01bd3a))
+* github cli v6 artifactName ([abbf1ab](https://github.com/nonodev96/THUMDER/commit/abbf1ab))
+* init repo ([1766be2](https://github.com/nonodev96/THUMDER/commit/1766be2))
 * Initialize angular10 ([c43f2d1](https://github.com/nonodev96/THUMDER/commit/c43f2d1))
+* new Workflow ([9c2c7f5](https://github.com/nonodev96/THUMDER/commit/9c2c7f5))
+* release fix? ([72bec50](https://github.com/nonodev96/THUMDER/commit/72bec50))
+* release github acctions ([75bfb16](https://github.com/nonodev96/THUMDER/commit/75bfb16))
+* Upload-release-asset ([73565a5](https://github.com/nonodev96/THUMDER/commit/73565a5))
+* Upload-release-asset ([b422a5c](https://github.com/nonodev96/THUMDER/commit/b422a5c))
+* workflow folder ([b9719a0](https://github.com/nonodev96/THUMDER/commit/b9719a0))